+++ notification.go
@@ -12,30 +12,30 @@ type NotificationsService struct {
 	client *Client
 }
 
-// NotificationDTO manages notifications
-type NotificationDTO struct {
+// Notification manages notifications
+type Notification struct {
 	Email       string                   `json:"email"`
 	PoolRecords []NotificationPoolRecord `json:"poolRecords"`
 }
 
 // NotificationPoolRecord does things unknown
 type NotificationPoolRecord struct {
-	PoolRecord   string              `json:"poolRecord"`
-	Notification NotificationInfoDTO `json:"notification"`
+	PoolRecord   string           `json:"poolRecord"`
+	Notification NotificationInfo `json:"notification"`
 }
 
-// NotificationInfoDTO does things unknown
-type NotificationInfoDTO struct {
+// NotificationInfo does things unknown
+type NotificationInfo struct {
 	Probe     bool `json:"probe"`
 	Record    bool `json:"record"`
 	Scheduled bool `json:"scheduled"`
 }
 
-// NotificationListDTO does things unknown
-type NotificationListDTO struct {
-	Notifications []NotificationDTO `json:"notifications"`
-	Queryinfo     QueryInfo         `json:"queryInfo"`
-	Resultinfo    ResultInfo        `json:"resultInfo"`
+// NotificationList does things unknown
+type NotificationList struct {
+	Notifications []Notification `json:"notifications"`
+	Queryinfo     QueryInfo      `json:"queryInfo"`
+	Resultinfo    ResultInfo     `json:"resultInfo"`
 }
 
 // NotificationKey collects the identifiers of an Notification
@@ -61,13 +61,13 @@ func (k NotificationKey) URI() string {
 }
 
 // Select requests all notifications by RRSetKey and optional query, using pagination and error handling
-func (s *NotificationsService) Select(k RRSetKey, query string) ([]NotificationDTO, *http.Response, error) {
+func (s *NotificationsService) Select(k RRSetKey, query string) ([]Notification, *http.Response, error) {
 	// TODO: Sane Configuration for timeouts / retries
 	maxerrs := 5
 	waittime := 5 * time.Second
 
 	// init accumulators
-	pis := []NotificationDTO{}
+	pis := []Notification{}
 	errcnt := 0
 	offset := 0
 
@@ -97,14 +97,14 @@ func (s *NotificationsService) Select(k RRSetKey, query string) ([]NotificationD
 }
 
 // SelectWithOffset requests list of notifications by RRSetKey, query and offset, also returning list metadata, the actual response, or an error
-func (s *NotificationsService) SelectWithOffset(k RRSetKey, query string, offset int) ([]NotificationDTO, ResultInfo, *http.Response, error) {
-	var tld NotificationListDTO
+func (s *NotificationsService) SelectWithOffset(k RRSetKey, query string, offset int) ([]Notification, ResultInfo, *http.Response, error) {
+	var tld NotificationList
 
 	uri := k.NotificationsQueryURI(query, offset)
 	res, err := s.client.get(uri, &tld)
 
 	log.Printf("DEBUG - ResultInfo: %+v\n", tld.Resultinfo)
-	pis := []NotificationDTO{}
+	pis := []Notification{}
 	for _, pi := range tld.Notifications {
 		pis = append(pis, pi)
 	}
@@ -112,19 +112,19 @@ func (s *NotificationsService) SelectWithOffset(k RRSetKey, query string, offset
 }
 
 // Find requests a notification by NotificationKey,returning the actual response, or an error
-func (s *NotificationsService) Find(k NotificationKey) (NotificationDTO, *http.Response, error) {
-	var t NotificationDTO
+func (s *NotificationsService) Find(k NotificationKey) (Notification, *http.Response, error) {
+	var t Notification
 	res, err := s.client.get(k.URI(), &t)
 	return t, res, err
 }
 
-// Create requests creation of an event by RRSetKey, with provided NotificationInfoDTO, returning actual response or an error
-func (s *NotificationsService) Create(k NotificationKey, n NotificationDTO) (*http.Response, error) {
+// Create requests creation of an event by RRSetKey, with provided NotificationInfo, returning actual response or an error
+func (s *NotificationsService) Create(k NotificationKey, n Notification) (*http.Response, error) {
 	return s.client.post(k.URI(), n, nil)
 }
 
-// Update requests update of an event by NotificationKey, with provided NotificationInfoDTO, returning the actual response or an error
-func (s *NotificationsService) Update(k NotificationKey, n NotificationDTO) (*http.Response, error) {
+// Update requests update of an event by NotificationKey, with provided NotificationInfoO, returning the actual response or an error
+func (s *NotificationsService) Update(k NotificationKey, n Notification) (*http.Response, error) {
 	return s.client.put(k.URI(), n, nil)
 }