+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.3.1] - 2017-03-03
 ### Changed
+- IPAddrRange: rename from IPAddrDTO
 - Client.NewRequest: shallow-copy BaseURL to avoid retaining modifications
 
 ## [1.3.0] - 2017-02-28